+ `--offset` command line argument
allows you to offset the timestamp used for comparison. this is used in the workflow - since it can't use "last_video_timestamp" as it can't save files (and therefore, defaults to the current time).  since it runs every two hours, it is now offset by 7200 seconds - allowing it to accept videos 7200 seconds before it was run.